requirements.txt
discord.py
asyncio
requests
shortuuid
